backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
pymongo[srv]==4.10.1
requests==2.32.3
rapidfuzz==3.9.6
pdfplumber==0.11.4
# Optional NLP enhancement (requires C++ toolchain on Windows; skip if build fails)
# spacy==3.7.5
google-generativeai==0.8.3
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
bcrypt==4.0.1
# Optional: download model with: python -m spacy download en_core_web_sm
